+++ src/components/shared/Container.vue
@@ -1,5 +1,0 @@
-<template>
-  <div class="w-full max-w-screen mx-auto px-4">
-    <slot />
-  </div>
-</template>