+++ src/main.ts
@@ -1,5 +1,5 @@
 import 'uno.css';
-import '@/style.css';
+import '@/style.scss';
 
 import { createApp } from 'vue';
 import App from '@/App.vue';
@@ -9,8 +9,8 @@ import { router } from '@/router';
 import { pinia } from '@/pinia';
 
 // Element Plus
-import ElementPlus from 'element-plus'
-import 'element-plus/dist/index.css'
+import ElementPlus from 'element-plus';
+import 'element-plus/dist/index.css';
 
 // Ipc
 import { initIpcLog } from './plugins/ipc/initIpcLog';