+++ src/plugins/ipc/initIpcLog.ts
@@ -8,12 +8,12 @@ export const initIpcLog = () => {
   }
 
   window.ipcRenderer.on('main-process-message', (_event, message) => {
-    console.log(message)
+    console.log(message);
   });
   window.ipcRenderer.on('main-process-error', (_event, message) => {
-    console.error(message)
+    console.error(message);
   });
   window.ipcRenderer.on('main-process-warn', (_event, message) => {
-    console.warn(message)
+    console.warn(message);
   });
 };