+++ vite.config.ts
@@ -1,10 +1,10 @@
-import path from 'node:path'
+import path from 'node:path';
 
-import { defineConfig } from 'vite'
+import { defineConfig } from 'vite';
 
-import vue from '@vitejs/plugin-vue'
-import UnoCSS from 'unocss/vite'
-import electron from 'vite-plugin-electron/simple'
+import vue from '@vitejs/plugin-vue';
+import UnoCSS from 'unocss/vite';
+import electron from 'vite-plugin-electron/simple';
 
 import dotenv from 'dotenv';
 
@@ -25,24 +25,26 @@ export default defineConfig({
     UnoCSS(),
 
     // Use electron plugin only when need to wrap the Electron app.
-    needElectronWrapper && electron({
-      main: {
-        // Shortcut of `build.lib.entry`.
-        entry: 'electron/main.ts',
-      },
-      preload: {
-        // Shortcut of `build.rollupOptions.input`.
-        // Preload scripts may contain Web assets, so use the `build.rollupOptions.input` instead `build.lib.entry`.
-        input: path.join(__dirname, 'electron/preload.ts'),
-      },
-      // Ployfill the Electron and Node.js API for Renderer process.
-      // If you want use Node.js in Renderer process, the `nodeIntegration` needs to be enabled in the Main process.
-      // See 👉 https://github.com/electron-vite/vite-plugin-electron-renderer
-      renderer: process.env.NODE_ENV === 'test'
-        // https://github.com/electron-vite/vite-plugin-electron-renderer/issues/78#issuecomment-2053600808
-        ? undefined
-        : {},
-    }),
+    needElectronWrapper &&
+      electron({
+        main: {
+          // Shortcut of `build.lib.entry`.
+          entry: 'electron/main.ts',
+        },
+        preload: {
+          // Shortcut of `build.rollupOptions.input`.
+          // Preload scripts may contain Web assets, so use the `build.rollupOptions.input` instead `build.lib.entry`.
+          input: path.join(__dirname, 'electron/preload.ts'),
+        },
+        // Ployfill the Electron and Node.js API for Renderer process.
+        // If you want use Node.js in Renderer process, the `nodeIntegration` needs to be enabled in the Main process.
+        // See 👉 https://github.com/electron-vite/vite-plugin-electron-renderer
+        renderer:
+          process.env.NODE_ENV === 'test'
+            ? // https://github.com/electron-vite/vite-plugin-electron-renderer/issues/78#issuecomment-2053600808
+              undefined
+            : {},
+      }),
   ],
 
   resolve: {
@@ -50,4 +52,4 @@ export default defineConfig({
       '@': path.resolve(__dirname, 'src'),
     },
   },
-})
+});